+++ src/routes/(pages)/mission/+page.js
@@ -1,0 +1,7 @@
+export function load() {
+	return {
+		title: 'The five marks of mission',
+		caption: 'Image by reenablack from Pixabay',
+		static: true,
+	};
+}